Separated Weather from City, decoupling the immediate weather feature.Mmade Forecast object with basic logic for extended results in new activity
