+++ app/src/main/java/com/daniyalh/WeatherWiseApp/presentation/WeatherController.java
@@ -1,44 +1,77 @@
 package com.daniyalh.WeatherWiseApp.presentation;
 
-import android.util.Log;
-import android.widget.Toast;
-
 import com.daniyalh.WeatherWiseApp.logic.FavouritesManager;
-import com.daniyalh.WeatherWiseApp.logic.IWeatherCallback;
 import com.daniyalh.WeatherWiseApp.logic.IWeatherManager;
 import com.daniyalh.WeatherWiseApp.logic.WeatherManager;
-import com.daniyalh.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
 import com.daniyalh.WeatherWiseApp.objects.City;
+import com.daniyalh.WeatherWiseApp.objects.CityWeather;
+import com.daniyalh.WeatherWiseApp.objects.Forecast;
+import com.daniyalh.WeatherWiseApp.objects.Weather;
 
 public class WeatherController {
     private static final String TAG = "WeatherController";
+    private static WeatherController instance;
 
-    private final WeatherManager weatherManager;
-    private final FavouritesManager favouritesManager;
-    private final ForecastDetailActivity context;
+    private WeatherManager weatherManager;
+    private FavouritesManager favouritesManager;
+    private WeatherDetailActivity context;
 
+    private CityWeather cityWeather;
     private City city;
+    private Weather weather;
 
-    public WeatherController(WeatherManager weatherManager, FavouritesManager favouritesManager, ForecastDetailActivity context) {
+    private WeatherController() {}
+
+    public static WeatherController getInstance() {
+        // singleton
+        if (instance == null) {
+            instance = new WeatherController();
+        }
+        return instance;
+    }
+
+    public void injectDependencies(WeatherManager weatherManager,
+                                   FavouritesManager favouritesManager,
+                                   WeatherDetailActivity context) {
         this.weatherManager = weatherManager;
-        //this.cityManager = cityManager;
         this.favouritesManager = favouritesManager;
         this.context = context;
     }
 
-    public void fetchWeather(int cityID, String cityName, String country, String country_code) {
+    public void fetchWeather(int cityID, String cityName, String country, String countryCode) {
+        cityWeather = new CityWeather(cityID);
+        cityWeather.setCountry(cityName, country, countryCode);
+
         city = new City(cityID);
-        city.setCountry(cityName, country, country_code);
+        city.setDetails(cityName, country, countryCode);
 
         weatherManager.getWeatherFromDB(city, new IWeatherManager.IWeatherDetailsCallback() {
             @Override
             public void onSuccess(String[] weatherDetails) {
                 context.showLoadingIcon(false);
-                city.updateWeather(weatherDetails);
+                weather = new Weather(weatherDetails);
+                city.setWeather(weather);
                 context.setCityLabel(city);
+                context.setStaticUIVisibility(true);
+                context.updateWeatherDetails(weather);
+            }
+
+            @Override
+            public void onError(String error) {
+
+            }
+        });
+
+        /*
+        weatherManager.getWeatherFromDB(cityWeather, new IWeatherManager.IWeatherDetailsCallback() {
+            @Override
+            public void onSuccess(String[] weatherDetails) {
+                context.showLoadingIcon(false);
+                cityWeather.updateWeather(weatherDetails);
+                context.setCityLabel(cityWeather);
 
                 context.setStaticUIVisibility(true);
-                context.updateWeatherDetails(city.getWeather());
+                context.updateWeatherDetails(cityWeather.getWeather());
             }
 
             @Override
@@ -46,41 +79,17 @@ public class WeatherController {
                 context.showLoadingIcon(false);
                 context.showAlertDialog("Error", "Unable to fetch weather data." +
                         " Please check the city name and try again.");
-                Log.e(TAG, "Error fetching weather for city " + city.getCity() + " - " + error);
+                Log.e(TAG, "Error fetching weather for city " + cityWeather.getCity() + " - " + error);
                 context.resetUI();
             }
-        });
-        /*
-        weatherManager.getWeatherJSON(city, new IWeatherManager.IWeatherCallback() {
-            @Override
-            public void onSuccess(String response) {
-                try {
-                    context.showLoadingIcon(false);
-                    weatherManager.setWeather(city, response);
-                    context.setCityLabel(city);
+        });*/
+    }
 
-                    context.setStaticUIVisibility(true);
-                    context.updateWeatherDetails(city.getWeather());
-
-                    //cityManager.addCity(city); // Add or update record
-                } catch (InvalidJsonParsingException e) {
-                    context.showToast(e.getMessage(), Toast.LENGTH_LONG);
-                }
-            }
-
-            @Override
-            public void onError(String error) {
-                context.showLoadingIcon(false);
-                context.showAlertDialog("Error", "Unable to fetch weather data." +
-                        " Please check the city name and try again.");
-                Log.e(TAG, "Error fetching weather for city " + city.getCity() + " - " + error);
-                context.resetUI();
-            }
-        });
-         */
+    public void fetchExtendedWeather() {
+        Forecast[] forecasts = new Forecast[40];
     }
 
     public void toggleFavourite(boolean isFavourite) {
-        favouritesManager.toggleFavourite(city.getCityID(), isFavourite);
+        favouritesManager.toggleFavourite(cityWeather.getCityID(), isFavourite);
     }
 }
