+++ app/src/main/res/layout/activity_weather_detail.xml
@@ -204,6 +204,19 @@
         app:lottie_loop="false"
         app:lottie_rawRes="@raw/favourite_icon" />
 
+    <com.airbnb.lottie.LottieAnimationView
+        android:id="@+id/extended_forecast_icon_lottie"
+        android:layout_width="44dp"
+        android:layout_height="44dp"
+        android:layout_marginEnd="16dp"
+        android:visibility="invisible"
+        app:layout_constraintBottom_toBottomOf="@+id/clear_favourites_button"
+        app:layout_constraintEnd_toStartOf="@+id/clear_favourites_button"
+        app:layout_constraintTop_toTopOf="@+id/clear_favourites_button"
+        app:layout_constraintVertical_bias="0.0"
+        app:lottie_loop="true"
+        app:lottie_rawRes="@raw/extended_forecast_icon" />
+
     <TextView
         android:id="@+id/wind_label_text_view"
         android:layout_width="wrap_content"
