+++ app/src/main/res/raw/extended_forecast_icon.json
@@ -1,0 +1,1 @@
+{"nm":"Main Scene","ddd":0,"h":300,"w":300,"meta":{"g":"@lottiefiles/creator 1.42.1"},"layers":[{"ty":4,"nm":"Layer 4","sr":1,"st":0,"op":180,"ip":0,"hd":false,"ddd":0,"bm":0,"hasMask":false,"ao":0,"ks":{"a":{"a":0,"k":[0,0,0],"ix":1},"s":{"a":0,"k":[2133.333,2133.333,100],"ix":6},"sk":{"a":0,"k":0},"p":{"a":1,"k":[{"o":{"x":0.001,"y":0},"i":{"x":0.58,"y":1},"s":[149.914,135.57],"t":14},{"o":{"x":0.001,"y":0},"i":{"x":0.58,"y":1},"s":[149.914,209.57],"t":25},{"o":{"x":0.167,"y":0.167},"i":{"x":0.833,"y":0.833},"s":[149.914,135.57],"t":40},{"o":{"x":0.001,"y":0},"i":{"x":0.58,"y":1},"s":[149.914,135.57],"t":62},{"o":{"x":0.001,"y":0},"i":{"x":0.58,"y":1},"s":[149.914,209.57],"t":73},{"o":{"x":0.167,"y":0},"i":{"x":0.58,"y":1},"s":[149.914,135.57],"t":86},{"o":{"x":0.001,"y":0},"i":{"x":0.58,"y":1},"s":[149.914,209.57],"t":98},{"s":[149.914,135.57],"t":113}],"ix":2},"r":{"a":0,"k":0,"ix":10},"sa":{"a":0,"k":0},"o":{"a":0,"k":100,"ix":11}},"shapes":[{"ty":"gr","bm":0,"hd":false,"mn":"ADBE Vector Group","nm":"Group 1","ix":1,"cix":2,"np":2,"it":[{"ty":"sh","bm":0,"hd":false,"mn":"ADBE Vector Shape - Group","nm":"Path 1","ix":1,"d":1,"ks":{"a":0,"k":{"c":false,"i":[[0,0],[0,0],[0,0]],"o":[[0,0],[0,0],[0,0]],"v":[[-5.3,-2.644],[-0.004,2.652],[5.3,-2.652]]},"ix":2}},{"ty":"st","bm":0,"hd":false,"mn":"ADBE Vector Graphic - Stroke","nm":"Stroke 1","lc":1,"lj":1,"ml":10,"o":{"a":0,"k":100,"ix":4},"w":{"a":0,"k":2,"ix":5},"c":{"a":0,"k":[0,0,0],"ix":3}},{"ty":"tr","a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"sk":{"a":0,"k":0,"ix":4},"p":{"a":0,"k":[0,0],"ix":2},"r":{"a":0,"k":0,"ix":6},"sa":{"a":0,"k":0,"ix":5},"o":{"a":0,"k":100,"ix":7}}]}],"ind":1}],"v":"5.7.0","fr":60,"op":180,"ip":0,"assets":[],"markers":[{"cm":"1","tm":84,"dr":0}]}
